feat(IQNet-App/client/src/views/Administrator):added delete city validations

+++ client/src/views/administrator/Administrator.jsx
@@ -79,7 +79,35 @@ const Administrator = () => {
   };
 
   const handleDeleteCity = (id) => {
-    dispatch(deleteCity(id));
+    Swal.fire({
+      title:
+        "Esta seguro que desea eliminar esta localidad del formulario de prospectos? Esto no se podra deshacer!",
+      showDenyButton: true,
+      confirmButtonText: "Confirmar",
+      denyButtonText: `Cancelar`,
+      color: "#5a5a5a",
+      confirmButtonColor: "#59A0FD",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire({
+          title: "Esta seguro?",
+          showDenyButton: true,
+          confirmButtonText: "Confirmar",
+          denyButtonText: `Cancelar`,
+          color: "#5a5a5a",
+          confirmButtonColor: "#59A0FD",
+        }).then((result) => {
+          if (result.isConfirmed) {
+            Swal.fire("Localidad Eliminada", "", "success");
+            dispatch(deleteCity(id));
+          } else if (result.isDenied) {
+            return;
+          }
+        });
+      } else if (result.isDenied) {
+        return;
+      }
+    });
   };
 
   const handleDeleteUser = (id) => {
